@@ -2,7 +2,7 @@
 #
 # Website: https://nylo.dev
 # Official repository: https://github.com/nylo-core/nylo
-# Author: Anthony Gordon <agordon@nylo.dev>
+# Author: Anthony Gordon <https://github.com/agordn52>
 
 name: flutter_app
 description: A new Nylo Flutter application.
@@ -12,7 +12,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.14.0 <3.0.0'
+  sdk: '>=2.15.0 <3.0.0'
 
 dependencies:
   url_launcher: ^6.0.17
@@ -21,7 +21,7 @@ dependencies:
   dio: ^4.0.4
   intl: ^0.17.0
   page_transition: ^2.0.4
-  nylo_framework: ^2.1.0
+  nylo_framework: ^2.1.1
   flutter:
     sdk: flutter
   flutter_localizations:
